# Improve pickling/unpickling performance 