--- a/dotnet/Razorvine.Pyrolite/Pyrolite/Pickle/UnpickleStack.cs
+++ b/dotnet/Razorvine.Pyrolite/Pyrolite/Pickle/UnpickleStack.cs
@@ -1,5 +1,6 @@
 /* part of Pyrolite, by Irmen de Jong (irmen@razorvine.net) */
 
+using System;
 using System.Collections;
 using System.Diagnostics.CodeAnalysis;
 
@@ -11,31 +12,42 @@ namespace Razorvine.Pickle
 /// </summary>
 [SuppressMessage("ReSharper", "InconsistentNaming")]
 public class UnpickleStack {
-	private readonly ArrayList _stack;
 	public readonly object MARKER;
+    private const int DefaultCapacity = 4;
+	private object[] _stack;
+    private int _count;
 
 	public UnpickleStack() {
-		_stack = new ArrayList();
+		_stack = new object[DefaultCapacity];
 		MARKER = new object(); // any new unique object
 	}
 
-	public void add(object o) {
-		_stack.Add(o);
-	}
+    public void add(object o)
+    {
+        if (_count == _stack.Length)
+        {
+            Array.Resize(ref _stack, _count * 2);
+        }
+
+        _stack[_count++] = o;
+    }
 
 	public void add_mark() {
-		_stack.Add(MARKER);
+		add(MARKER);
 	}
 
-	public object pop() {
-		int size = _stack.Count;
-		var result = _stack[size - 1];
-		_stack.RemoveAt(size-1);
-		return result;
-	}
+    public object pop()
+    {
+        if (_count == 0)
+        {
+            throw new ArgumentOutOfRangeException("index"); // match exception type/parameter name thrown from ArrayList used previously
+        }
+
+        return _stack[--_count];
+    }
 
 	public ArrayList pop_all_since_marker() {
-		ArrayList result = new ArrayList();
+		var result = new ArrayList();
 		object o = pop();
 		while (o != MARKER) {
 			result.Add(o);
@@ -46,21 +58,39 @@ public class UnpickleStack {
 		return result;
 	}
 
+    internal object[] pop_all_since_marker_as_array()
+    {
+        int i = _count - 1;
+        while (_stack[i] != MARKER)
+            i--;
+
+        var result = new object[_count - 1 - i];
+        Array.Copy(_stack, i + 1, result, 0, result.Length);
+
+        _count = i;
+        return result;
+    }
+
 	public object peek() {
-		return _stack[_stack.Count-1];
-	}
+        return _stack[_count - 1];
+    }
 
 	public void trim() {
-		_stack.TrimToSize();
+        if (_count < _stack.Length && _stack.Length > DefaultCapacity)
+        {
+            var newArr = new object[Math.Max(_count, DefaultCapacity)];
+            Array.Copy(_stack, 0, newArr, 0, _count);
+            _stack = newArr;
+        }
 	}
 
 	public int size() {
-		return _stack.Count;
+		return _count;
 	}
 
 	public void clear() {
-		_stack.Clear();
-		_stack.TrimToSize();
+        _count = 0;
+		trim();
 	}
 }