Thanks for the helpful library!

I found while using it that it was generating a bunch of intermediate objects that could be avoided, and that these were contributing non-trivially to execution time.  This PR addresses many of them while also avoiding touching changing the public surface area, as I wasn't sure if it was malleable or not.  There's more that can be done in this regard, but for now this addresses a lot of the low-hanging fruit that's there.

(I also disabled one test that was failing prior to any of my changes.)

Example:
```C#
using System.IO;
using System.Linq;
using Razorvine.Pickle;
using BenchmarkDotNet.Attributes;
using BenchmarkDotNet.Running;

[MemoryDiagnoser]
public class Test
{
    public static void Main() => BenchmarkRunner.Run<Test>();

    private readonly MemoryStream _stream = new MemoryStream();

    private object[][] _objectArrays = Enumerable.Range(0, 1000).Select(_ => new object[] { 1.0, 2.0 }).ToArray();
    private double[][] _doubleArrays = Enumerable.Range(0, 1000).Select(_ => new double[] { 1.0, 2.0 }).ToArray();
    private string[][] _stringArrays = Enumerable.Range(0, 1000).Select(_ => new string[] { "hello", "world" }).ToArray();

    [Benchmark] public void ObjectArrays() => PickleUnpickle(_objectArrays);
    [Benchmark] public void DoubleArrays() => PickleUnpickle(_doubleArrays);
    [Benchmark] public void StringArrays() => PickleUnpickle(_stringArrays);

    private void PickleUnpickle(object obj)
    {
        _stream.SetLength(0);
        new Pickler(useMemo: false).dump(obj, _stream);
        _stream.Position = 0;
        new Unpickler().load(_stream);
    }
}
```

Before:
```
|       Method |       Mean |     Error |   StdDev |    Gen 0 |   Gen 1 | Gen 2 |  Allocated |
|------------- |-----------:|----------:|---------:|---------:|--------:|------:|-----------:|
| ObjectArrays |   444.1 us |  9.490 us | 21.81 us |  78.6133 | 11.7188 |     - |  321.79 KB |
| DoubleArrays | 1,250.1 us | 26.046 us | 32.94 us | 312.5000 |       - |     - | 1282.73 KB |
| StringArrays |   633.2 us | 11.740 us | 12.06 us | 124.0234 |  1.9531 |     - |  509.29 KB |
```

After:
```
|       Method |     Mean |    Error |   StdDev |   Gen 0 |  Gen 1 | Gen 2 | Allocated |
|------------- |---------:|---------:|---------:|--------:|-------:|------:|----------:|
| ObjectArrays | 234.0 us | 2.081 us | 1.946 us | 26.8555 | 4.1504 |     - | 110.31 KB |
| DoubleArrays | 660.7 us | 5.529 us | 5.172 us | 89.8438 | 0.9766 |     - |  368.4 KB |
| StringArrays | 415.5 us | 3.102 us | 2.901 us | 61.0352 | 0.9766 |     - | 250.94 KB |
```

cc: @imback82, @eerhardt, @rapoth