@eerhardt 2019-04-16
@stephentoub - Do you think it would be valuable to add your benchmark test to the repo?