@stephentoub 2019-04-16
> Do you think it would be valuable to add your benchmark test to the repo?

That would be up to @irmen. There's currently no perf tests project, and the only perf test anywhere for the C# functionality is minimal and disabled (https://github.com/irmen/Pyrolite/blob/87de7710ae6b3a6b36b869f9d797feecbefa19aa/dotnet/Razorvine.Pyrolite/Tests/Pickle/UnpicklerTests.cs#L515-L516).  Adding perf tests is only valuable if someone will actually be using them to either monitor for regressions or drive further improvements.