@irmen 2019-04-16
@stephentoub may I ask in general, what you are using Pyrolite for?