@stephentoub 2019-04-16
> may I ask in general, what you are using Pyrolite for?

Yup.  Been looking at using it as part of https://issues.apache.org/jira/browse/SPARK-27006.