@stephentoub 2019-04-18
@irmen, please let me know if there's anything else you'd like me to address.  I'm happy to make whatever changes would help to get this in and into an updated NuGet.