@irmen 2019-04-19
just published the new version to nuget.org as well.