@irmen 2019-04-19
thanks for the improvements and interest in pyrolite, really appreciated